Working nb integration test, at least for C* 4.0.  Previous test framework was generalized a bit and converted into a pair of mixins which can be used to implement tests of this type very quickly. Hopefully the mixins can generalize in the future if need be.
While testing I also hit upon a discrepency between how nb YAML was being generated in the output-dir
case vs. the base stdout case.  Added an explicit strip() to the stdout case to avoid the extra empty
lines that appear to be added by the "yaml" module.  I don't love this solution... but it does seem
to work (at least so far).